requirements.txt
cs50
flask
flask_session 
requests
datetime
tempfile
werkzeug.exceptions
werkzeug.security